requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alpaca-trade-api==3.2.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backtrader==1.9.78.123
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
deprecation==2.1.0
fastapi==0.115.12
Flask==3.1.0
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.0.3
multidict==6.3.2
multitasking==0.0.11
numpy==2.2.4
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
yarl==1.19.0
yfinance==0.2.55
